* changes:
  fix(optee): initialize the structure
  fix(optee): add missing curly braces
  fix(optee): add parenthesis for macro expressions
  fix(optee): move function to conditional block
  fix(optee): add boolean type for expressions
  fix(optee): evaluate condition for boolean
  fix(optee): typecast operands to match data type
  fix(el3-runtime): resolve essential-type mismatch

+++ plat/xilinx/versal/sip_svc_setup.c
@@ -122,6 +122,6 @@ DECLARE_RT_SVC(
 		sip_svc,
 		OEN_SIP_START,
 		OEN_SIP_END,
-		(uint8_t)SMC_TYPE_FAST,
+		SMC_TYPE_FAST,
 		sip_svc_setup,
 		sip_svc_smc_handler);
